prepare clean portable build without warnings

+++ src/build_log_perftest.cc
@@ -122,7 +122,7 @@ int main() {
   int max = times[0];
   float total = 0;
   for (size_t i = 0; i < times.size(); ++i) {
-    total += times[i];
+    total += static_cast<float>(times[i]);
     if (times[i] < min)
       min = times[i];
     else if (times[i] > max)
@@ -130,7 +130,7 @@ int main() {
   }
 
   printf("min %dms  max %dms  avg %.1fms\n",
-         min, max, total / times.size());
+         min, max, total / static_cast<float>(times.size()));
 
   unlink(kTestFilename);
